requirements.txt
Flask>=2.0
Pillow>=8.0
numpy>=1.19
tensorflow>=2.0
